requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
billiard==4.2.0
blinker==1.7.0
cachetools==5.3.2
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.10
django-redis==5.4.0
facebook-sdk==3.1.0
flask==3.0.2
flower==2.0.1
frozenlist==1.4.1
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
humanize==4.9.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
MarkupSafe==2.1.5
multidict==6.0.5
oauthlib==3.2.2
pillow==10.2.0
prometheus-client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.1
six==1.16.0
sqlparse==0.4.4
tornado==6.4
typing-extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
